refactor(register/login): move files to pages folder and adjust components
