+++ FinEdgeFrontend/src/Pages/Register.tsx
@@ -2,44 +2,27 @@ import { useState, SyntheticEvent, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import axios, { AxiosError } from "axios";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { Button } from "@progress/kendo-react-buttons";
-import {
-  TextBox,
-  TextBoxChangeEvent,
-  InputSuffix,
-} from "@progress/kendo-react-inputs";
-import { Tooltip } from "@progress/kendo-react-tooltip";
+import { Button, TextBox, TextBoxChangeEvent, InputSuffix, Tooltip } from "@progress/kendo-react-all";
 import { SvgIcon } from "@progress/kendo-react-common";
 import "@progress/kendo-theme-default/dist/default-ocean-blue.css";
-import {
-  envelopeIcon,
-  userIcon,
-  passwordIcon,
-} from "@progress/kendo-svg-icons";
-import { registerEndPoint } from "../../endpoints";
-import { RegisterDTO } from "../../Utils/Types";
-import { isValidPassword, isValidEmail } from "../../Utils/Functions";
-import Questions from "../Questions/Questions";
-import CustomLink from "../../Components/CustomLink";
+import { envelopeIcon, userIcon, passwordIcon } from "@progress/kendo-svg-icons";
+import { registerEndPoint } from "../endpoints";
+import { RegisterDTO } from "../Utils/Types";
+import { isValidPassword, isValidEmail } from "../Utils/Functions";
+import Questions from "../Components/HomePage/Questions";
+import CustomLink from "../Components/CustomLink";
 
 const Register = () => {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-  const emailRegEx = new RegExp(
-    "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
-  );
-  const passwordRegEx = new RegExp(
-    "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{10,}$"
-  );
+  const emailRegEx = new RegExp("^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$");
+  const passwordRegEx = new RegExp("^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{10,}$");
 
   const [emailError, setEmailError] = useState<boolean>(true);
   const [passwordError, setPasswordError] = useState<boolean>(true);
-  const [isNextButtonDisabled, setIsNextButtonDisabled] =
-    useState<boolean>(true);
-  const [isNextButtonClicked, setIsNextButtonClicked] =
-    useState<boolean>(false);
-  const [isRegisterButtonDisabled, setIsRegisterButtonDisabled] =
-    useState<boolean>(true);
+  const [isNextButtonDisabled, setIsNextButtonDisabled] = useState<boolean>(true);
+  const [isNextButtonClicked, setIsNextButtonClicked] =useState<boolean>(false);
+  const [isRegisterButtonDisabled, setIsRegisterButtonDisabled] = useState<boolean>(true);
   const [user, setUser] = useState<RegisterDTO>({
     name: "",
     surname: "",
@@ -113,20 +96,10 @@ const Register = () => {
           </div>
 
           <form onSubmit={handleSubmit} method="post" autoComplete="off">
-            <div
-              className={`space-y-6 ${
-                isNextButtonClicked ? "hidden" : "block"
-              }`}
-            >
+            <div className={`space-y-6 ${isNextButtonClicked ? "hidden" : "block"}`}>
               <div className="space-y-6">
                 <div className="mb-2">
-                  <TextBox
-                    id="name"
-                    type="text"
-                    name="name"
-                    placeholder="Name"
-                    onChange={handleChange}
-                    required={true}
+                  <TextBox id="name" type="text" name="name" placeholder="Name" onChange={handleChange} required={true}
                     prefix={() => (
                       <InputSuffix>
                         <SvgIcon icon={userIcon} />
@@ -136,13 +109,7 @@ const Register = () => {
                 </div>
 
                 <div className="mb-2">
-                  <TextBox
-                    id="surname"
-                    type="text"
-                    name="surname"
-                    placeholder="Surname"
-                    onChange={handleChange}
-                    required={true}
+                  <TextBox id="surname" type="text" name="surname" placeholder="Surname" onChange={handleChange} required={true}
                     prefix={() => (
                       <InputSuffix>
                         <SvgIcon icon={userIcon} />
@@ -152,13 +119,7 @@ const Register = () => {
                 </div>
 
                 <div className="mb-2">
-                  <TextBox
-                    id="email"
-                    type="email"
-                    name="email"
-                    placeholder="Email"
-                    onChange={handleChange}
-                    required={true}
+                  <TextBox id="email" type="email" name="email" placeholder="Email" onChange={handleChange} required={true}
                     prefix={() => (
                       <InputSuffix>
                         <SvgIcon icon={envelopeIcon} />
@@ -169,13 +130,7 @@ const Register = () => {
 
                 <div className="mb-2">
                   <Tooltip anchorElement="target">
-                    <TextBox
-                      id="password"
-                      type="password"
-                      name="password"
-                      placeholder="Password"
-                      onChange={handleChange}
-                      required={true}
+                    <TextBox id="password" type="password" name="password" placeholder="Password" onChange={handleChange} required={true} 
                       title="Password needs to be at least 10 characters, have 1 upper/lower case letter, 1 number and 1 special symbol"
                       prefix={() => (
                         <InputSuffix>
@@ -188,25 +143,14 @@ const Register = () => {
               </div>
 
               <div className="text-center">
-                <Button
-                  id="next"
-                  themeColor={"primary"}
-                  disabled={isNextButtonDisabled}
-                  onClick={handleClick}
-                  size={"large"}
-                  className="w-64 font-bold .text-2xl"
-                >
+                <Button id="next" themeColor={"primary"} disabled={isNextButtonDisabled} onClick={handleClick} size={"large"} className="w-64 font-bold text-2xl">
                   Next
                 </Button>
               </div>
 
               <div className="mt-6 text-center text-sm text-gray-600">
-                <p>
-                  Have an account?{" "}
-                  <CustomLink
-                    to="/login"
-                    className="text-blue-600 hover:text-blue-700 font-medium"
-                  >
+                <p>Have an account?{" "}
+                  <CustomLink to="/login" className="text-blue-600 hover:text-blue-700 font-medium">
                     Log in
                   </CustomLink>
                 </p>
@@ -214,33 +158,14 @@ const Register = () => {
             </div>
 
             <div
-              className={`${
-                isNextButtonClicked ? "block" : "hidden"
-              } space-y-8`}
-            >
+              className={`${isNextButtonClicked ? "block" : "hidden"} space-y-8`}>
               <div className="bg-blue-100 text-black px-5 py-3 rounded-lg">
                 <Questions onMethodologyChange={handleMethodologyChange} />
               </div>
 
               <div className="flex space-x-3 justify-center">
-                <Button
-                  id="backButton"
-                  className="w-48 font-bold"
-                  themeColor={"error"}
-                  size={"large"}
-                  onClick={handleClick}
-                >
-                  Back
-                </Button>
-                <Button
-                  id="registerButton"
-                  className="w-48 font-bold"
-                  themeColor={"primary"}
-                  size={"large"}
-                  disabled={isRegisterButtonDisabled}
-                >
-                  Register
-                </Button>
+                <Button id="backButton" className="w-48 font-bold" themeColor={"error"} size={"large"} onClick={handleClick}>Back</Button>
+                <Button id="registerButton" className="w-48 font-bold" themeColor={"primary"} size={"large"} disabled={isRegisterButtonDisabled}>Register</Button>
               </div>
             </div>
           </form>
